# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Event Management API",
    "description": "This is a backend for event management application",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/api"
    }
  ],
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "user"
        ],
        "operationId": "HandleCreateUser",
        "summary": "create user",
        "description": "**Requires authenticated user with the 'admin' role.**",
        "requestBody": {
          "description": "user's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful user created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RestAPIResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/UnauthorizedResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "user is missing the required role 'admin'"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ServerErrorResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "repository connection lost"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "user"
        ],
        "operationId": "HandleGetUserById",
        "summary": "get user by id",
        "description": "**Requires authenticated user with the 'admin' role.**",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuidv4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RestAPIResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/UnauthorizedResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "user is missing the required role 'admin'"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ServerErrorResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "repository connection lost"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "user"
        ],
        "operationId": "HandleUpdateUserById",
        "summary": "update user by id",
        "description": "**Requires authenticated user with the 'admin' role.**",
        "requestBody": {
          "description": "user's information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrUpdateUser"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuidv4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RestAPIResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BadRequestResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/UnauthorizedResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "user is missing the required role 'admin'"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ServerErrorResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "repository connection lost"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "user"
        ],
        "operationId": "HandleDeleteUserById",
        "summary": "delete user by id",
        "description": "**Requires authenticated user with the 'admin' role.**",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuidv4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/RestAPIResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/BadRequestResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/UnauthorizedResponse"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Error",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ServerErrorResponse"
                    },
                    {
                      "properties": {
                        "messages": {
                          "example": "repository connection lost"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "tokenAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RestAPIResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "ErrorAPIResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/RestAPIResponse"
          },
          {
            "properties": {
              "success": {
                "example": false
              },
              "code": {
                "type": "string",
                "example": "AUTH_INVALID_CREDENTIALS"
              },
              "messages": {
                "oneOf": [
                  {
                    "type": "string",
                    "example": "more information about the error..."
                  },
                  {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "more information about the error..."
                    }
                  }
                ]
              }
            }
          }
        ]
      },
      "BadRequestResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "type": "string",
                "example": "BAD_REQUEST"
              },
              "messages": {
                "example": [
                  "validation error"
                ]
              }
            }
          }
        ]
      },
      "ServerErrorResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "example": "INTERNAL_SERVER_ERROR"
              },
              "messages": {
                "example": [
                  "Service is unavailable"
                ]
              }
            }
          }
        ]
      },
      "UnauthorizedResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ErrorAPIResponse"
          },
          {
            "properties": {
              "code": {
                "example": "AUTH_INVALID_CREDENTIALS"
              },
              "messages": {
                "example": [
                  "invalid credentials"
                ]
              }
            }
          }
        ]
      },
...
      "CreateOrUpdateUser": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/UserData"
          },
          {
            "type": "object",
            "properties": {
              "role": {
                "type": "string",
                "enum": [
                  "user",
                  "organizer",
                  "admin"
                ]
              },
              "verified": {
                "type": "boolean",
                "example": true
              }
            }
          }
        ]
      },
      "UserData": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "curiousdev@bedevs.org"
          },
          "first_name": {
            "type": "string",
            "example": "john"
          },
          "last_name": {
            "type": "string",
            "example": "doe"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "e5f04d06-a6ff-4522-a5f4-86b3c150991c"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "example": "curiousdev@bedevs.org"
          },
          "first_name": {
            "type": "string",
            "example": "john"
          },
          "second_name": {
            "type": "string",
            "example": "doe"
          },
          "birthday": {
            "$ref": "#/components/schemas/NullTime"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "organizer",
              "admin"
            ]
          },
          "verified": {
            "type": "boolean",
            "example": false
          },
          "about": {
            "$ref": "#/components/schemas/NullString"
          },
          "created_at": {
            "type": "string",
            "example": "2024-06-14T19:17:50.776Z",
            "format": "datetime"
          },
          "updated_at": {
            "type": "string",
            "example": "2024-06-14T19:17:50.776Z",
            "format": "datetime"
          }
        }
      },
      "NullTime": {
        "type": "object",
        "properties": {
          "Time": {
            "type": "string"
          },
          "Valid": {
            "type": "boolean"
          }
        }
      },
      "NullString": {
        "type": "object",
        "properties": {
          "String": {
            "type": "string"
          },
          "Valid": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "tokenAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}